requirements.txt
pymongo
requests
python-dotenv
fastapi
uvicorn
httpx
pytest
beautifulsoup4
selenium
